requirements.txt
allure-pytest==2.9.43
apipkg==1.5
attrs==21.2.0
beautifulsoup4==4.9.3
execnet==1.8.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-xdist==2.2.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.5
